Added Login Functionality
You can now login as long as you have a user in the database. If the
user has a temp password it takes them to the screen to change it, and
you can succesfully change your password. If the user is a student and
does not have a temp password it takes them to the student main page, if
they are an admin it takes them to the admin_overall page.